Don't pipe data into curl for healthchecks
This doesn't actually work for healthchecks. Healthchecks just gets the start of the request and assumes everything is working. If the process dies later, healthchecks just sees that as receiving no more data, so assumes everything is fine, when it's not.

This way we lose the logs from healthchecks, but gain it actually working. It can be got back with a wrapper, but that's more than 1 line.